% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafdR_utils.R
\name{gen.rnd.vec}
\alias{gen.rnd.vec}
\title{Generates random vector for numerical answer (internal)}
\usage{
gen.rnd.vec()
}
\value{
A vector with random weights, including value 1 for first element
}
\description{
Generates random vector for numerical answer (internal)
}
\examples{
library(pafdR)
my.sol <- 1 #assume solution to exercise is 1

 my.answers <- gen.rnd.vec()*my.sol
 my.answers
}
