% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafdR_utils.R
\name{pafdR_check.pdflatex}
\alias{pafdR_check.pdflatex}
\title{Function to check if system has pdflatex.exe available}
\usage{
pafdR_check.pdflatex()
}
\value{
TRUE if the pdflatex is available, FALSE if not
}
\description{
Function to check if system has pdflatex.exe available
}
\examples{
pafdR_check.pdflatex()

}
