% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafdR_download_code_data.R
\name{pafdR_download.code.and.data}
\alias{pafdR_download.code.and.data}
\title{Downloads code and data from the book}
\usage{
pafdR_download.code.and.data(lan = "en")
}
\arguments{
\item{lan}{Language of book/files ('en' or 'pt-br')}
}
\value{
TRUE, if successful
}
\description{
Downloads a zip file from the github repository and unzip it to a folder within
the working directory
}
\examples{

\dontrun{
pafdR_download.code.and.data()
}
}
