% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pafdR_download_exercises.R
\name{pafdR_download.exercises}
\alias{pafdR_download.exercises}
\title{Downloads Book exercises from github}
\usage{
pafdR_download.exercises(exercise.folder = "pafdR-exercises")
}
\arguments{
\item{exercise.folder}{folder to copy exercise files}
}
\value{
TRUE, if sucessful
}
\description{
This function downloads exercise files from github and unzips it to a chosen directory
}
\examples{

\dontrun{
exercise.folder <- pafdR_download.exercises()
}

}
